Rename loop labels to avoid a warning
There are two separate loops called 'digits present in the same
function. Although these are on two different sides of an if/else
statement, meaning that they can never overlap, there was a language
change several months ago that forbade even this in the name of
future-proofing:

- https://github.com/rust-lang/rust/pull/24162
- https://internals.rust-lang.org/t/psa-rejecting-duplicate-loop-labels/1833

This commit merely gives them unique names, removing the warning.